Added QOL Improvements to the library
