+++ .github/workflows/test.yml
@@ -1,0 +1,39 @@
+name: Test
+
+on:
+  pull_request:
+  push:
+    branches:
+      - main
+    tags:
+      - '*'
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+      - name: Setup Miniconda
+        uses: conda-incubator/setup-miniconda@v3
+        with:
+          auto-update-conda: true
+          auto-activate-base: false
+      - name: Install dependencies
+        shell: bash -l {0}
+        run: |
+          make conda-install
+      - name: Lint code with pre-commit
+        shell: bash -l {0}
+        run: |
+          make lint
+      - name: Install mcl with pip
+        shell: bash -l {0}
+        run: |
+          make build
+      - name: Run pytest
+        shell: bash -l {0}
+        run: |
+          make test
