+++ MAKEFILE
@@ -1,0 +1,17 @@
+.PHONY: build test docs env all
+all:
+	make lint && make build && make test
+build:
+	python -m pip install -vv -e .
+test:
+	pytest -vv
+lint:
+	pre-commit run --verbose --all-files
+docs:
+	cd docs && make html
+conda-env:
+	conda create -n mcl
+conda-install:
+	conda install python=3.12 pytest pre-commit numpy
+clean:
+	git clean -dfX
